v2.0-20200621
Update to support Kali Linux Version 2019 or higher (Solution to Kernel Panic)

+++ rotate.sh
@@ -165,7 +165,11 @@ fi
 sudo sync
 sudo sync
 
-echo "reboot now"
 sleep 1
-sudo reboot
-
+read -r -p "Reboot is required, are you sure? " response
+if [[ "$response" =~ ^([yY][eE][sS]|[yY])$ ]]
+then
+    sudo reboot
+else
+    echo "Restart manually to apply the changes."
+fi
